Refactors the Modal to add Tasks
Previously, the modal to add tasks was represented by a single component.
This commit separates it into two components: TaskModal [with the form inputs],
and TaskModalWrapper, which contains the TaskModal and a button to trigger the
TaskModal to open. This change is needed because previously, the modal can only be
opened by one trigger. This change allows external components to trigger the
modal to open.