Merge List#handleUpdate and List#editTask methods
Previously, #handleUpdate was used to mark tasks as completed
and #editTask was used to edit the information of the task.
As both methods performed the same function of updating tasks,
the #handleUpdate method was removed, and #editTask was renamed
to #handleUpdate. All user actions to edit tasks or mark tasks
as completed will be processed with the new #handleUpdate method.