Ensures consistency among date displays
Previously, the date format will change to the ISO format
when the task's deadline is edited. This is because a new Date object is stored
when updating the deadline, and that Date object is converted
into a date string with the ISO format.

This commit ensures all displayed dates follow the "yyyy-mm-dd"
format.