Clear form information after adding tasks
Previously, information entered to add a task would persist
in the Task Modal after adding a task. This commit fixes
the issue by conditionally rendering the Task Modal.
As a new Task Modal component is created when a task is added,
the state of the Task Modal is reset each time.