Refactor: Daily and OneOffList to ConnectedList
Previously DailyList and OneOffList were two different list components.
But they were very similar and introducing changes in list components meant
I had to duplicate a lot of code written and increased the possibilities of errors.

The solution is to create a generic component, ConnectedList, which subscribes
to the Redux store.